Update: Media library, reqest media function
New "Request Media" Workflow:

In the designer's portal, when viewing a "Scheduled" post, a new "Request Media" button will appear.

Clicking this button will change the post's status to "Awaiting Media Upload" and trigger a notification for the client.

A new "Awaiting Media Upload" column will be added to the designer's "Overview" so these posts are easy to track.

Client Media Upload Portal:

When a post is in the "Awaiting Media Upload" status, clients will see a new file uploader within the post's detail view.

This will allow clients to upload images, videos, and other document types directly to the post. This upload functionality is only available when media has been explicitly requested.

Once the client uploads the files, the post status will automatically revert to "Scheduled," signaling to the designer that the assets are ready.

Central Media Library:

A new "Media Library" tab will be added to the main navigation for both clients and designers.

This section will display all media that has been uploaded to any post.

The files will be organized and grouped by the post they belong to, creating an intuitive and centralized place to find all assets.