Update to status
I will update the application with the following adjustments:

Status Renamed: All instances of the "Scheduled" status will be changed to "In Progress" throughout the application.

Client Portal Update: The "In Progress" status will now be visible on the client's side.

Updated "Overview": A new "In Progress" column will be added to the "Overview" tab for both the designer and the client, ensuring everyone can see posts that are being prepared.